backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
psycopg2-binary>=2.9
python-dotenv>=1.0
pandas>=2.0
numpy>=1.26
scikit-learn>=1.4
joblib>=1.3
